Guild joined and count folder move
